==> Scripts/0_Data_Download/9_Download_Pathway_alterations.R <==
### Pathway alterations data was obtained from the supplementary data of 
### https://www.cell.com/cell/fulltext/S0092-8674(18)30359-3?_returnURL=https%3A%2F%2Flinkinghub.elsevier.com%2Fretrieve%2Fpii%2FS0092867418303593%3Fshowall%3Dtrue
